requirements.txt
qrcode-7.3.1
pyzbar-0.1.9
pillow
